+++ database/watcher.ts
@@ -5,14 +5,14 @@ const { Client } = pg
 console.log('YO WATCHER')
 
 const listenConnectionString = process.env.LISTEN_DATABASE_URL!
-const writeConnectionString = process.env.WRITE_DATABASE_URL!
+const authConnectionString = process.env.WRITE_DATABASE_URL!
 
 export const listenClient = new Client({
   connectionString: listenConnectionString,
 })
 
 export const authDb = new Client({
-  connectionString: writeConnectionString,
+  connectionString: authConnectionString,
 })
 
 listenClient
@@ -49,7 +49,7 @@ async function ensureTablesExist() {
         userid TEXT PRIMARY KEY,
         "to" TEXT,
         recovery TEXT,
-        timestamp TIMESTAMP,
+        timestamp TEXT,
         log_addr TEXT,
         block_num NUMERIC
       )
@@ -61,8 +61,8 @@ async function ensureTablesExist() {
         id TEXT PRIMARY KEY,
         userid TEXT NOT NULL REFERENCES public.users(userid),
         deviceid TEXT NOT NULL,
-        created TIMESTAMP,
-        expiresAt TIMESTAMP NOT NULL
+        created TEXT,
+        expiresAt TEXT NOT NULL
       )
     `)
 
@@ -74,6 +74,7 @@ async function ensureTablesExist() {
         deviceid TEXT NOT NULL,
         publickey TEXT NOT NULL,
         encryptedprivatekey TEXT NOT NULL,
+        timestamp TEXT NOT NULL
         PRIMARY KEY (userid, custodyAddress, deviceid)
       )
     `)
@@ -110,7 +111,7 @@ async function checkAndReplicateData() {
       const res = await authDb.query(
         `
   INSERT INTO public.users (userid, "to", recovery, timestamp, log_addr, block_num)
-  SELECT * FROM unnest($1::NUMERIC[], $2::TEXT[], $3::TEXT[], $4::TIMESTAMP[], $5::TEXT[], $6::NUMERIC[])
+  SELECT * FROM unnest($1::NUMERIC[], $2::TEXT[], $3::TEXT[], $4::TEXT[], $5::TEXT[], $6::NUMERIC[])
   ON CONFLICT (userid) DO UPDATE
   SET "to" = EXCLUDED."to", recovery = EXCLUDED.recovery, timestamp = EXCLUDED.timestamp, log_addr = EXCLUDED.log_addr, block_num = EXCLUDED.block_num
   RETURNING *
