+++ routes/signArbitrary.ts
